+++ src/components/Organisms/Modals/Searches/SkillSearchModal.vue
@@ -2,8 +2,11 @@
 import Vue from 'vue';
 import { API_URL } from '@/master'
 import axios from 'axios'
-import { Skill, FetchSkills } from '@/types/index';
-import { Job } from '@/types/job';
+import { 
+  Job,
+  Skill,
+} from '@/types/index';
+import { FetchSkills } from '@/types/fetch';
 
 type DateType = {
   skills: Skill[];
