+++ src/views/Jobs/JobDetail.vue
@@ -13,7 +13,7 @@ import PostUser from '@/components/Organisms/Jobs/JobDetails/PostUser.vue'
 import SkillJob from '@/components/Organisms/Jobs/JobDetails/SkillJob.vue'
 import DetailJob from '@/components/Organisms/Jobs/JobDetails/DetailJob.vue'
 import BtnArea from '@/components/Organisms/Jobs/JobDetails/BtnArea.vue'
-import { FetchJobs } from '@/types/job';
+import { FetchJobs } from '@/types/fetch';
 
 type State = {
   job: any;
