+++ src/views/Manages/Applications/Apply.vue
@@ -8,10 +8,8 @@ import {
 } from '@vue/composition-api';
 import { API_URL, m } from '@/master'
 import axios from 'axios'
-import { 
-  ManageJob,
-  FetchManageJobs
-} from '@/types/manage';
+import { ManageJob } from '@/types/index';
+import { FetchManageJobs } from '@/types/fetch';
 import UserCard from '@/components/Organisms/Manages/UserCard.vue'
 import JobsCard from '@/components/Organisms/Manages/JobsCard.vue'
 import Vuex from '@/store/index'
@@ -89,7 +87,7 @@ export default defineComponent({
           </v-row>
           <v-col>
             <router-link 
-              :to="`/manage/apply_job/${ jobs.jobId }`" 
+              :to="`/manage/apply_job/${ jobs.job_id }`" 
               v-for="jobs in applyJob" 
               :key="jobs.id" 
               class="jobs"
