+++ src/views/Users/Manages/ManageUserProfile.vue
@@ -12,7 +12,7 @@ import PostUser from '@/components/Organisms/Users/PostUser.vue'
 import SkillUser from '@/components/Organisms/Users/SkillUser.vue'
 import IntroduceUser from '@/components/Organisms/Users/IntroduceUser.vue'
 import StatusChangeBtnArea from '@/components/Organisms/Manages/StatusChangeBtnArea.vue'
-import { User } from '@/types/user';
+import { User } from '@/types/index';
 import { API_URL, truncate } from '@/master'
 // import Logout from '@/components/button/Logout'
